1.6

+++ Parent/RealtyFrontend/src/main/java/org/realty/DelUserCommand.java
@@ -6,20 +6,19 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.realty.UserJdbcDAO;
 
 public class DelUserCommand implements Command {
-	
-	
+
 	public String execute(HttpServletRequest request,
 			HttpServletResponse response) throws ServletException, IOException {
-		
 
-                UserJdbcDAO ad = new UserJdbcDAO();
+		UserJdbcDAO ad = new UserJdbcDAO();
 		String delUserId = request.getParameter("delUserId");
-		ad.delete(Long.parseLong(delUserId));
+		ad.delete((Long)Long.parseLong(delUserId));
 
 		return CommandFactory.getCommand("allUser").execute(request, response);
-		
+
 	}
 
 }
